+++ src/compiler/frontend.c
@@ -238,7 +238,7 @@ static TAB_table userLabels=NULL; // Temp_label->TRUE, line numbers, explicit la
  *
  *******************************************************************/
 
-#define MAX_KEYWORDS 94
+#define MAX_KEYWORDS 95
 
 S_symbol FE_keywords[MAX_KEYWORDS];
 int FE_num_keywords;
@@ -337,6 +337,7 @@ static S_symbol S_ACCESS;
 static S_symbol S_CLOSE;
 static S_symbol S_BREAK;
 static S_symbol S__FNO;
+static S_symbol S_DPRINT;
 
 static inline bool isSym(S_tkn tkn, S_symbol sym)
 {
@@ -2988,7 +2989,7 @@ static bool stmtExternDecl(S_tkn *tkn, E_enventry e, CG_item *exp)
 }
 
 // print ::= PRINT [ # expFNo , ] [ expression ( [ ';' | ',' ] expression )* ]
-static bool stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
+static bool _stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp, bool dbg)
 {
     S_pos pos = (*tkn)->pos;
     *tkn = (*tkn)->next; // skip "PRINT"
@@ -2996,6 +2997,8 @@ static bool stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
     CG_item exFNo;
     if ((*tkn)->kind == S_HASH)
     {
+        if (dbg)
+            return EM_error(pos, "DPRINT: fno expression not supported.");
         *tkn = (*tkn)->next;
         S_pos pos = (*tkn)->pos;
         if (!expression(tkn, &exFNo))
@@ -3019,9 +3022,12 @@ static bool stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
 
         CG_itemList arglist = CG_ItemList();
         CG_itemListNode n;
-        n = CG_itemListAppend(arglist);
-        n->item = exFNo;
-        CG_loadVal (g_sleStack->code, pos, &n->item);
+        if (!dbg)
+        {
+            n = CG_itemListAppend(arglist);
+            n->item = exFNo;
+            CG_loadVal (g_sleStack->code, pos, &n->item);
+        }
         n = CG_itemListAppend(arglist);
         n->item = ex;
         CG_loadVal (g_sleStack->code, pos, &n->item);
@@ -3030,34 +3036,34 @@ static bool stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
         switch (ty->kind)
         {
             case Ty_string:
-                fsym = S_Symbol("_aio_puts", FALSE);
+                fsym = dbg ? S_Symbol("_debug_puts", FALSE) : S_Symbol("_aio_puts", FALSE);
                 break;
             case Ty_pointer:
-                fsym = S_Symbol("_aio_putu4", FALSE);
+                fsym = dbg ? S_Symbol("_debug_putu4", FALSE) :  S_Symbol("_aio_putu4", FALSE);
                 break;
             case Ty_byte:
-                fsym = S_Symbol("_aio_puts1", FALSE);
+                fsym = dbg ? S_Symbol("_debug_puts1", FALSE) :  S_Symbol("_aio_puts1", FALSE);
                 break;
             case Ty_ubyte:
-                fsym = S_Symbol("_aio_putu1", FALSE);
+                fsym = dbg ? S_Symbol("_debug_putu1", FALSE) :  S_Symbol("_aio_putu1", FALSE);
                 break;
             case Ty_integer:
-                fsym = S_Symbol("_aio_puts2", FALSE);
+                fsym = dbg ? S_Symbol("_debug_puts2", FALSE) :  S_Symbol("_aio_puts2", FALSE);
                 break;
             case Ty_uinteger:
-                fsym = S_Symbol("_aio_putu2", FALSE);
+                fsym = dbg ? S_Symbol("_debug_putu2", FALSE) :  S_Symbol("_aio_putu2", FALSE);
                 break;
             case Ty_long:
-                fsym = S_Symbol("_aio_puts4", FALSE);
+                fsym = dbg ? S_Symbol("_debug_puts4", FALSE) :  S_Symbol("_aio_puts4", FALSE);
                 break;
             case Ty_ulong:
-                fsym = S_Symbol("_aio_putu4", FALSE);
+                fsym = dbg ? S_Symbol("_debug_putu4", FALSE) :  S_Symbol("_aio_putu4", FALSE);
                 break;
             case Ty_single:
-                fsym = S_Symbol("_aio_putf", FALSE);
+                fsym = dbg ? S_Symbol("_debug_putf", FALSE) :  S_Symbol("_aio_putf", FALSE);
                 break;
             case Ty_bool:
-                fsym = S_Symbol("_aio_putbool", FALSE);
+                fsym = dbg ? S_Symbol("_debug_putbool", FALSE) :  S_Symbol("_aio_putbool", FALSE);
                 break;
             default:
                 return EM_error(pos, "unsupported type in print expression list.");
@@ -3085,7 +3091,7 @@ static bool stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
             case S_COMMA:
             {
                 *tkn = (*tkn)->next;
-                S_symbol fsym   = S_Symbol("_aio_puttab", FALSE);
+                S_symbol fsym   =  dbg ? S_Symbol("_debug_puttab", FALSE) : S_Symbol("_aio_puttab", FALSE);
                 E_enventryList lx = E_resolveSub(g_sleStack->env, fsym);
                 if (!lx)
                     return EM_error(pos, "builtin %s not found.", S_name(fsym));
@@ -3108,7 +3114,7 @@ static bool stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
 
     if (isLogicalEOL(*tkn))
     {
-        S_symbol fsym   = S_Symbol("_aio_putnl", FALSE);
+        S_symbol fsym   = dbg ? S_Symbol("_debug_putnl", FALSE) : S_Symbol("_aio_putnl", FALSE);
         E_enventryList lx = E_resolveSub(g_sleStack->env, fsym);
         if (!lx)
             return EM_error(pos, "builtin %s not found.", S_name(fsym));
@@ -3123,6 +3129,15 @@ static bool stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
     }
 
     return FALSE;
+}
+
+static bool stmtPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
+{
+    return _stmtPrint (tkn, e, exp, /*dbg=*/FALSE);
+}
+static bool stmtDPrint(S_tkn *tkn, E_enventry e, CG_item *exp)
+{
+    return _stmtPrint (tkn, e, exp, /*dbg=*/TRUE);
 }
 
 static bool inputVar(S_tkn *tkn)
@@ -7106,6 +7121,7 @@ static void registerBuiltins(void)
     declareBuiltinProc(S_LBOUND       , /*extraSyms=*/ NULL      , funLBound        , Ty_ULong());
     declareBuiltinProc(S_UBOUND       , /*extraSyms=*/ NULL      , funUBound        , Ty_ULong());
     declareBuiltinProc(S__ISNULL      , /*extraSyms=*/ NULL      , funIsNull        , Ty_Bool());
+    declareBuiltinProc(S_DPRINT       , /*extraSyms=*/ NULL      , stmtDPrint       , Ty_Void());
 }
 
 //
@@ -7503,6 +7519,7 @@ void FE_boot(void)
     S_CLOSE           = defineKeyword("CLOSE");
     S_BREAK           = defineKeyword("BREAK");
     S__FNO            = defineKeyword("_FNO");
+    S_DPRINT          = defineKeyword("DPRINT");
 }
 
 void FE_init(void)